fix: move helper functions to provider, remove the helper class and refactor tests
